# Conflicts:
#	tests/unit/Mage/Catalog/Model/Resource/Eav/AttributeTest.php
#	tests/unit/Mage/Customer/Model/CustomerTest.php
#	tests/unit/Mage/Reports/Helper/DataTest.php
#	tests/unit/Mage/Reports/Model/Resource/Report/CollectionTest.php
#	tests/unit/Mage/Rule/Model/AbstractTest.php
#	tests/unit/Mage/Rule/Model/EnvironmentTest.php
#	tests/unit/Mage/Rule/Model/RuleTest.php
#	tests/unit/Traits/DataProvider/Base/BoolTrait.php
#	tests/unit/Traits/DataProvider/Base/IntOrNullTrait.php
#	tests/unit/Traits/DataProvider/Base/ModulesTrait.php
#	tests/unit/Traits/DataProvider/Base/NumericStringTrait.php
#	tests/unit/Traits/DataProvider/Mage/Catalog/CatalogTrait.php
#	tests/unit/Traits/DataProvider/Mage/Catalog/Model/LocaleTrait.php
#	tests/unit/Traits/DataProvider/Mage/Cms/CmsTrait.php
#	tests/unit/Traits/DataProvider/Mage/Core/CoreTrait.php
#	tests/unit/Traits/DataProvider/Mage/Downloadable/DownloadableTrait.php
#	tests/unit/Traits/DataProvider/Mage/Reports/ReportsTrait.php
#	tests/unit/Traits/DataProvider/Mage/Rule/RuleTrait.php
#	tests/unit/Traits/DataProvider/Mage/Uploader/UploaderTrait.php

+++ app/code/core/Mage/Catalog/Model/Resource/Category/Indexer/Product.php
@@ -956,9 +956,15 @@ class Mage_Catalog_Model_Resource_Category_Indexer_Product extends Mage_Index_Mo
              * Clean up temporary tables
              */
             $this->clearTemporaryIndexTable();
-            $idxAdapter->delete($enabledTable);
-            $idxAdapter->delete($anchorTable);
-            $idxAdapter->delete($anchorProductsTable);
+            if (isset($enabledTable)) {
+                $idxAdapter->delete($enabledTable);
+            }
+            if (isset($anchorTable)) {
+                $idxAdapter->delete($anchorTable);
+            }
+            if (isset($anchorProductsTable)) {
+                $idxAdapter->delete($anchorProductsTable);
+            }
             $this->commit();
         } catch (Exception $e) {
             $this->rollBack();
